+++ tests/test_untrack_command.rs
@@ -38,6 +38,17 @@ fn test_untrack() {
     std::fs::write(repo_path.join(".gitignore"), "*.bak\n").unwrap();
     let files_before = test_env.jj_cmd_success(&repo_path, &["files"]);
 
+    // Errors out when no path is specified
+    let stderr = test_env.jj_cmd_failure(&repo_path, &["untrack"]);
+    insta::assert_snapshot!(stderr.replace("jj.exe", "jj"), @r###"
+    error: The following required arguments were not provided:
+        <PATHS>...
+
+    USAGE:
+        jj untrack [OPTIONS] <PATHS>...
+
+    For more information try --help
+    "###);
     // Errors out when a specified file is not ignored
     let stderr = test_env.jj_cmd_failure(&repo_path, &["untrack", "file1", "file1.bak"]);
     insta::assert_snapshot!(stderr, @"Error: 'file1' would be added back because it's not ignored. Make sure it's ignored, \